Fluid Mesh Collision + several other features 0. we implement fluid mesh collision detection using height map generated on-the-fly. 1. add a shader inspector window for displaying program interface 2. use indexed drawing instead of triangle soup 3. Implement instance rendering (for fun) 4. scene monitor for acquiring info of the objects in the current scene 5. shader inspector window to query shader program interface/resources. 6. primitive drawing for debugging 7. texture mapping support 8. shader program binaries support (avoid recompiling the shaders each time the program runs)

+++ XRealityRender/src/simulation/SPHfluid/SpatialGrid.h
@@ -4,13 +4,13 @@
 #include <vector>
 #include "SPHCommon.h"
 
-/* 
- * SpatialGrid discretizes the space into equally-spaced grids
- * each grid stores the indeices of the particle in order to
- * speed up the neighborhood particle finding. 
- *
- * Che-Yuan Liang Apr. 2016 
- */
+/*
+* SpatialGrid discretizes the space into equally-spaced grids
+* each grid stores the indeices of the particle in order to
+* speed up the neighborhood particle finding.
+*
+* Che-Yuan Liang Apr. 2016
+*/
 
 
 class SpatialGrid
@@ -41,14 +41,14 @@ protected:
 		const double py = p.position.y;
 		const double pz = p.position.z;
 
-		return (int)(floor((px-base_x)/grid_h) + floor((py-base_y)/grid_h) * dim_x + floor((pz-base_z)/grid_h) * dim_x * dim_y);
+		return floor((px - base_x) / grid_h) + floor((py - base_y) / grid_h) * dim_x + floor((pz - base_z) / grid_h) * dim_x * dim_y;
 	}
 
 public:
 	SpatialGrid(double base_x, double base_y, double base_z,
-				double length_x, double length_y, double length_z, double h, double r);
+		double length_x, double length_y, double length_z, double h, double r);
 	~SpatialGrid();
-	
+
 	void addParticle(const SPHSim::SPHParticle &p, int i);
 	void clearGrid();
 	void findNeighbors(SPHSim::SPHParticle &p_i, const std::vector<SPHSim::SPHParticle> &particles);
